pyproject.toml
[project]
name = "rag-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.0",
    "pydantic-settings>=2.10.1",
    "pypdf>=5.7.0",
    "python-docx>=1.2.0",
    "uvicorn>=0.35.0",
]
